requirements.txt
pydantic
pyyaml
pandas
requests
fastapi
uvicorn
pytest
numpy-financial
httpx
